requirements.txt
aiohttp==3.10.9
aiosignal==1.2.0
async-timeout==4.0.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==2.1.1
discord.py==2.0.1
idna==3.4
requests==2.31.0
soupsieve==2.5
urllib3==2.0.6
paho-mqtt==2.1.0
audioop-lts==0.2.1
pytz
av
Pillow